fix($ngVueProvider): support passing arbitrary properties to root Vue instance.
Sometimes we need to inject properties in the root Vue instance (i.e., Vuex's `store`,
Vue i18n `i18n`, Vue Router's `router`) and the current structure of `ngVue` doesn't allow
us to do that. This change is a proposal to give the user of the library freedom to pass
any property to the root vue instance created by `ngVue` so that even custom plugins may
be used with it.

In order to keep away from a breaking change, a warning is issued when `$ngVueProvider.enableVuex()` is called. A
new method `$ngVueProvider.setRootViewInstanceProps()` was added where the Vuex store can be added. It
also explicitly _forbids_ passing Vue hooks as a "root prop" as there is already a mechanism to merge
the components lifecycle hooks with `ngVue` callbacks.